refactor(backend): separate app and server concerns
Extract Express application setup into a dedicated app.js file to
improve separation of concerns. This change makes the codebase more
maintainable and testable by isolating the server configuration from
the application logic. The server.js file now only handles the server
startup process.
